Move Cmdline out of arch-specific runtime.
We would surely keep the same parsing between arches, with just some
per-arch limits on options.  Make the struct and parsing common.
